moved bluetooth code to merchprofile
